backend/pyproject.toml
[project]
name = "nueera-growth-os-backend"
version = "0.1.0"
description = "Nueera Growth OS - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.4",
    "uvicorn[standard]==0.30.6",
    "SQLAlchemy==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "python-jose==3.4.0",
    "passlib[bcrypt]==1.7.4",
    "email-validator==2.2.0",
    "python-multipart==0.0.22",
    "celery==5.4.0",
    "redis==5.0.8",
    "httpx==0.27.0",
    "rich==13.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]
